requirements.txt
pandas
streamlit
numpy
langchain
langchain-openai
python-dotenv
fpdf
PyPDF2
markdown-pdf
pydantic